requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
websockets==12.0
google-cloud-speech==2.24.0
aiokafka==0.10.0
pydantic==2.5.3
pydantic-settings==2.1.0
prometheus-client==0.19.0
structlog==24.1.0
tenacity==8.2.3
httpx==0.26.0
python-multipart==0.0.7
